Fix CI coverage comment permission
If PR is from a foreign repo, the presubmit workflow doesn't have enough
permission to comment on the PR for security reasons. See details at
https://securitylab.github.com/research/github-actions-preventing-pwn-requests/
so we move the coverage comment on PR to another workflow.

If lcov-reporter-action implements the feature to write to the job summary,
we can revert this change.